requirements.txt
streamlit==1.46.0
python-dotenv==1.0.0
openai==1.10.0
httpx==0.24.1
pandas>=2.2.0
openpyxl==3.1.2
